Change amounts arguments in Input, Transaction
Transaction.fee/amount will become possible to be equal None.
Input.amount will become an optional argument. This will be done in
order for the tx object to be more consistent with the original
signature of the tx on the bitcoin network (in it, the serialized tx
doesn't contain the input amount and fee).

Also add SupportsSerialize interface.